Refactor code to include getErrorMessage function
This commit refactors the code in the sendEmail module by including the getErrorMessage function in the helper module. This function is used to retrieve the error message in case of a failed email sending. It handles cases when the error is an instance of the Error class or when it is a string, object with a 'message' property, or an unknown value.